Added instructions to limit the nextcloud quota to 1 Byte
I tested this, works great. It only effects the nextcloud data, not the external shares, so I think we still need to have the file security rules I submitted earlier. I don't think we need to do anything on the Linux side to disable automatic syncing of data, which is nice.